fixed rebuild of Evolution trunk: install stamp must be removed, otherwise it won't reinstall
git-svn-id: https://zeitsenke.de/svn/SyncEvolution/trunk@519 15ad00c4-1369-45f4-8270-35d70d36bdcd

+++ test/runtests.py
@@ -500,7 +500,9 @@ class EvoSvn(Action):
         cd(self.workdir)
         shutil.copy2(os.path.join(self.makedir, "Makefile"), ".")
         shutil.copy2(os.path.join(self.makedir, "local.mk"), ".")
-        shutil.rmtree(self.resultdir)
+        if os.access(self.resultdir, os.F_OK):
+            shutil.rmtree(self.resultdir)
+        os.system("rm -f .stamp/*.install")
 	localmk = open("local.mk", "a")
 	localmk.write("PREFIX := %s\n" % self.resultdir)
 	localmk.close()
